Move all Board functions under one namespace: Board
