--- lib/board/manipulation.ex
@@ -1,7 +1,0 @@
-defmodule Board.Manipulation do
-
-  def add_move(board, {cell_index, players_mark}) do
-    List.replace_at(board, cell_index, players_mark)
-  end
-
-end
